Move the response data to `src/net/response/`
