+++ src/index.ts
@@ -17,13 +17,14 @@ export { checksum, validate } from './modules/utils/CRC16';
 export { DataPayload } from './modules/data/DataPayload';
 export { SodiumHelper } from './modules/utils/SodiumHelper';
 export { Hash, hash, hashMulti, makeUTXOKey, hashFull, hashPart } from './modules/data/Hash';
-export { PreImage  } from './modules/data/PreImage';
 export { KeyPair, PublicKey, SecretKey, Seed } from './modules/data/KeyPair';
 export { Signature } from './modules/data/Signature';
-export { Validator } from './modules/data/Validator';
 export { Transaction, TxType } from './modules/data/Transaction';
 export { TxInput } from './modules/data/TxInput';
 export { TxOutput } from './modules/data/TxOutput';
+
+export { Validator } from './modules/net/response/Validator';
+export { PreImage  } from './modules/net/response/PreImage';
 
 export { BOAClient } from './modules/net/BOAClient';
 export { Request } from './modules/net/Request';
