+++ src/modules/utils/index.ts
@@ -1,4 +1,3 @@
 export { checksum, validate } from './CRC16';
 export { readFromString, writeToString } from './buffer';
 export { SodiumHelper } from './SodiumHelper';
-export { validateJSON } from './json';
